Add a separate command with a separate permission to join a clan
